+++ wwooss-api/src/main/java/com/bigbro/wwooss/v1/service/es/TrashCanDocumentService.java
@@ -17,5 +17,5 @@ public interface TrashCanDocumentService {
      * trashType : 쓰레기 타입 -> ,로 구분 PLASTIC,CAN
      * distance : 내 기준 반경거리 단위 - KM
      */
-    List<TrashCanResponse> findTrashCanByGeoLocation(Double lat, Double lng, Integer distance, String trashType);
+    List<TrashCanResponse> findTrashCanByGeoLocation(Double lat, Double lng, Double distance, String trashType);
 }
