+++ wwooss-api/src/main/java/com/bigbro/wwooss/v1/service/es/impl/TrashCanDocumentServiceImpl.java
@@ -45,7 +45,7 @@ public class TrashCanDocumentServiceImpl implements TrashCanDocumentService {
 
     @Override
     @Transactional(readOnly = true)
-    public List<TrashCanResponse> findTrashCanByGeoLocation(Double lat, Double lng, Integer distance, String trashType) {
+    public List<TrashCanResponse> findTrashCanByGeoLocation(Double lat, Double lng, Double distance, String trashType) {
         List<TrashCanDocument> findTrashCanLocation = trashCanDocumentRepository.findByGeoLocationAndTrashType(lat, lng, distance, trashType);
         return findTrashCanLocation.stream().map(trashCan -> {
             String trashTypeListString = trashCan.getTrashType();
