+++ wwooss-domain-es/src/main/java/com/bigbro/wwooss/v1/repository/custom/TrashCanDocumentRepositoryCustom.java
@@ -8,5 +8,5 @@ import java.util.List;
 public interface TrashCanDocumentRepositoryCustom {
     void saveTrashCan(List<TrashCanInfo> trashCanInfoList);
 
-    List<TrashCanDocument> findByGeoLocationAndTrashType(Double lat, Double lng, Integer distance, String trashType);
+    List<TrashCanDocument> findByGeoLocationAndTrashType(Double lat, Double lng, Double distance, String trashType);
 }
